deep-learning/1-pixel-attack/requirements.txt
jupyter
keras
tensorflow-gpu
pandas
numpy
matplotlib
scipy
tqdm
requests